Created a cable icon as CAN bus indicator, smaller than the three letters CAN.

+++ main/IpsDisplay.h
@@ -106,10 +106,10 @@ private:
 	static void drawCircling( int x, int y, bool draw );
 	static void drawCruise( int x, int y, bool draw );
 	static void drawBT();
-	static void drawCable(int x, int y);
+	static void drawCable(int16_t x, int16_t y);
 	static void drawFlarm( int x, int y, bool flarm );
 	static void drawWifi( int x, int y );
-	static void drawConnection( int x, int y );
+	static void drawConnection( int16_t x, int16_t y );
 	static void drawBat( float volt, int x, int y, bool blank );
 	static void drawTemperature( int x, int y, float t );
 	static void drawThermometer( int x, int y );
